requirements.txt
python-telegram-bot == 13.15.0
tinydb == 4.5.2
django == 3.2.7